requirements.txt
# Updated Dependencies for Production LifeFlow.ai

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9  # PostgreSQL
asyncpg>=0.29.0  # Async PostgreSQL

# Authentication (existing)
pyjwt>=2.8.0
python-multipart>=0.0.6

# LLM & AI
openai>=1.6.0
anthropic>=0.8.0  # Optional: Claude fallback
tiktoken>=0.5.2  # Token counting

# Vector Database
faiss-cpu>=1.7.4  # Development
pinecone-client>=2.2.4  # Production (optional)
# milvus>=2.3.0  # Alternative vector DB

# Embeddings & Search
sentence-transformers>=2.2.2
rank-bm25>=0.2.2  # Hybrid search

# Document Processing
pypdf2>=3.0.1
pdfplumber>=0.10.3  # Better PDF extraction
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-docx>=1.1.0
markdown>=3.5.1

# Web Scraping
httpx>=0.25.2
playwright>=1.40.0  # For dynamic content
aiohttp>=3.9.1

# Background Tasks
celery>=5.3.4
redis>=5.0.1

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3  # Retry logic
numpy>=1.26.2
pandas>=2.1.4  # Data processing

# Monitoring & Logging
loguru>=0.7.2
sentry-sdk>=1.39.1  # Error tracking

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2  # For testing

# Development
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.1
